requirements.txt
fastapi
uvicorn[standard]
httpx
sqlalchemy
sqlalchemy-utils
databases[aiomysql]
python-dotenv
pydantic-settings
rich
asgi-correlation-id
python-json-logger
logtail-python
python-jose
python-multipart
passlib[bcrypt]
aiofiles
mysql-connector-python==8.3.0
boto3
alembic
cryptography